+++ asapdiscovery-ml/asapdiscovery/ml/__init__.py
@@ -7,4 +7,4 @@ from .es import BestEarlyStopping, ConvergedEarlyStopping  # noqa: F401
 from .inference import *  # noqa: F401,F403
 from .loss import *  # noqa: F401,F403
 from .pretrained_models import *  # noqa: F401,F403
-from .weights import *  # noqa: F401,F403
+from .models import *  # noqa: F401,F403
