+++ asapdiscovery-ml/asapdiscovery/ml/models/__init__.py
@@ -1,4 +1,4 @@
-from .weights import (
+from .ml_models import (
     ASAPMLModelRegistry,
     MLModelType,
     MLModelBase,
